Move file from editors folder
