+++ src/explorer/setAppWizardContextDefault.ts
@@ -8,8 +8,8 @@ import * as path from 'path';
 import { Uri, workspace, WorkspaceConfiguration } from 'vscode';
 import { IAppServiceWizardContext, LinuxRuntimes, WebsiteOS } from 'vscode-azureappservice';
 import { LocationListStep } from 'vscode-azureextensionui';
-import { configurationSettings, extensionPrefix } from '../../constants';
-import { javaUtils } from '../../utils/javaUtils';
+import { configurationSettings, extensionPrefix } from '../constants';
+import { javaUtils } from '../utils/javaUtils';
 
 export async function setAppWizardContextDefault(wizardContext: IAppServiceWizardContext): Promise<void> {
     const isJavaProject: boolean = await javaUtils.isJavaProject();
